+++ AppLinking/Editor/Settings/UnityCloudPlayerSettingsEditor.cs
@@ -84,20 +84,6 @@ namespace Unity.Cloud.AppLinking.Editor
         }
     }
 
-    [InitializeOnLoad]
-    internal class UnityCloudPlayerSettingsUpdater
-    {
-        static UnityCloudPlayerSettingsUpdater()
-        {
-            UnityCloudPlayerSettings.Instance.OnAppNamespaceChanged += OnAppNamespaceChanged;
-        }
-
-        static void OnAppNamespaceChanged()
-        {
-            EditorUtility.SetDirty(UnityCloudPlayerSettings.Instance);
-        }
-    }
-
     /// <summary>
     /// Draws a property as read-only in the GUI.
     /// </summary>