+++ AppLinking/Runtime/Settings/UnityCloudPlayerSettings.cs
@@ -129,23 +129,6 @@ namespace Unity.Cloud.AppLinking.Runtime
         {
             return Instance.AppNamespace;
         }
-
-#if UNITY_EDITOR
-        /// <summary>
-        /// Sets the app namespace.
-        /// </summary>
-        /// <remarks>Only available from Unity Editor scripts. Unlike the setter of the <see cref="AppNamespace"/> property,
-        /// using the <see cref="SetAppNamespace"/> method will update the associated .asset file on disk.
-        /// </remarks>
-        public void SetAppNamespace(string appNamespace)
-        {
-            Instance.AppNamespace = appNamespace;
-            OnAppNamespaceChanged?.Invoke();
-        }
-
-        public event Action OnAppNamespaceChanged;
-#endif
-
     }
 
     /// <summary>