+++ Core/Utilities/WebGL/TaskExtensions.cs
@@ -1,10 +1,13 @@
+using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 
 namespace Unity.Cloud.Common
 {
-    internal static class TaskExtensions
+    static class TaskExtensions
     {
-        internal static async Task<T> UnityConfigureAwait<T>(this Task<T> task, bool continueOnCapturedContext)
+        public static bool MultithreadingEnabled { get; set; } = true;
+
+        public static async Task<T> UnityConfigureAwait<T>(this Task<T> task, bool continueOnCapturedContext)
         {
 #if UNITY_WEBGL && !UNITY_EDITOR
             return await task;
@@ -13,13 +16,28 @@ namespace Unity.Cloud.Common
 #endif
         }
 
-        internal static async Task UnityConfigureAwait(this Task task, bool continueOnCapturedContext)
+        public static async Task UnityConfigureAwait(this Task task, bool continueOnCapturedContext)
         {
 #if UNITY_WEBGL && !UNITY_EDITOR
             await task;
 #else
             await task.ConfigureAwait(continueOnCapturedContext);
 #endif
+        }
+
+        public static ConfiguredTaskAwaitable ConfigureAwaitFalse(this Task task)
+        {
+            return task.ConfigureAwait(!MultithreadingEnabled);
+        }
+
+        public static ConfiguredValueTaskAwaitable ConfigureAwaitFalse(this ValueTask task)
+        {
+            return task.ConfigureAwait(!MultithreadingEnabled);
+        }
+
+        public static ConfiguredValueTaskAwaitable<T> ConfigureAwaitFalse<T>(this ValueTask<T> task)
+        {
+            return task.ConfigureAwait(!MultithreadingEnabled);
         }
     }
 }